refactor: singleton for file service
