+++ packages/neuron-wallet/src/services/wallets.ts
@@ -6,7 +6,7 @@ import Key, { Addresses } from '../keys/key'
 import { Keystore } from '../keys/keystore'
 import Store from '../utils/store'
 import { nodeService } from './node'
-import fileService from '../startup/file-service'
+import { fileService } from './file'
 import LockUtils from '../utils/lock-utils'
 import env from '../env'
 import i18n from '../utils/i18n'
