--- packages/neuron-wallet/src/startup/file-service.ts
@@ -1,3 +1,0 @@
-import FileService from '../services/file'
-
-export default new FileService()
